Rename YQPkgPackageKitGroupsFilterView -> YQPkgPackageClassificationFilterView
