+++ Dockerfile
@@ -1,4 +1,4 @@
 # Use the libyui/devel image as the base
-FROM libyui/devel
+FROM registry.opensuse.org/devel/libraries/libyui/containers/libyui-devel:latest
 
 COPY . /usr/src/app
