+++ SOURCECONF.cmake
@@ -26,7 +26,7 @@ SET( ${TARGETLIB}_SOURCES
   YQPkgLangList.cc
   YQPkgList.cc
   YQPkgObjList.cc
-  YQPkgPackageKitGroupsFilterView.cc
+  YQPkgPackageClassificationFilterView.cc
   YQPkgPatchFilterView.cc
   YQPkgPatchList.cc
   YQPkgPatternList.cc
@@ -75,7 +75,7 @@ SET( ${TARGETLIB}_HEADERS
   YQPkgLangList.h
   YQPkgList.h
   YQPkgObjList.h
-  YQPkgPackageKitGroupsFilterView.h
+  YQPkgPackageClassificationFilterView.h
   YQPkgPatchFilterView.h
   YQPkgPatchList.h
   YQPkgPatternList.h
